Remove suggestion of usage Debian Unstable repos on RPi
Adding Debian Unstable repositories and further system update effectively converts Debian Stable to Debian Unstable which is not recommended for production machines.

Prior to version 3.7, Thingsboard had dependency on Java 11 that was not available on Debian Stable/Bookworm (hovewer, it was temporally kept on repos of older and "unstable" Debian versions).
Thingsboard v3.7 depends on Java 17 LTS and the openjdk-17-jdk package  is available in the Debian Stable repositories.